Rewrite poseidon benchmarks with criterion
